feat(library): add real-time keyword suggestions with preset icons and library priority to folder editor
