+++ feature/library/README.md
@@ -9,7 +9,7 @@ Owns Library presentation: hub, history, subscriptions, liked episodes, download
 - `LibraryScreen` and `LibraryViewModel`. When Appearance **Cleaner Home** is on, the hub top bar shows Settings and Feedback (the same shortcuts Home normally owns).
 - `HistoryScreen` and `HistoryViewModel`. History uses a compact period-first hierarchy: a full-width listening-time hero, content-sized highlights, listening-pattern cards, then the filtered episode timeline. Highlights use a compact top-show row and explicit two-column metric grid instead of fixed-height carousel pages; paired cards match the tallest cell in their row, while an odd final metric spans the row as a featured card with a large tonal icon. The hero flows vertically so narrow phones do not sacrifice labels or leave split-column dead space; timeline rows allow longer episode/show copy. The one-time tracking-reset notice no longer appears.
 - `SubscriptionsScreen`, `LikedEpisodesScreen`, and `DownloadedEpisodesScreen`.
-- `FolderEditSheet` and `FolderEditComponents`: Subscription folder creation and edit sheet featuring a streamlined hierarchy, prominent name input with quick-fill suggestions, visual 3×3 grid footprint diagrams for 7 display sizes up to max 3×3 (`COMPACT 1×1`, `WIDE 2×1`, `FEATURED 2×2`, `LARGE 2×3`, `SHELF 3×1`, `PANEL 3×2`, `SHOWCASE 3×3`), contextual 1×1 cover display selector (`[ Folder Icon ]` vs `[ Podcast Grid ]`), compact optional icon picker row, LazyRow-powered horizontal scrolling with crisp edge stops (overscroll bounce disabled to prevent stretch distortion and floaty spring delay), and auto-sync with genre tags.
+- `FolderEditSheet` and `FolderEditComponents`: Subscription folder creation and edit sheet featuring a streamlined hierarchy, prominent name input with real-time keyword suggestions and preset icons (prioritizing subscribed library genres), visual 3×3 grid footprint diagrams for 7 display sizes up to max 3×3 (`COMPACT 1×1`, `WIDE 2×1`, `FEATURED 2×2`, `LARGE 2×3`, `SHELF 3×1`, `PANEL 3×2`, `SHOWCASE 3×3`), contextual 1×1 cover display selector (`[ Folder Icon ]` vs `[ Podcast Grid ]`), compact optional icon picker row, LazyRow-powered horizontal scrolling with crisp edge stops (overscroll bounce disabled to prevent stretch distortion and floaty spring delay), and auto-sync with genre tags.
 - Downloads multi-select: checklist in the top bar, or long-press a show (hub) / episode (show list) to enter selection with that row checked, then delete several at once.
 - `SmartDownloadsSettingsScreen` and `AutoDownloadSettingsScreen`.
 - `PlayAllFab` and library UI helpers.
